OHRM5X-1454: separated review summary into component
